move id to BaseElement
